+++ lib/brief/document/section/builder.rb
@@ -40,6 +40,7 @@ class Brief::Document::Section
       until even? || maxed_out?
         source.map! do |item|
           level, fragments = item
+          fragments = Array(fragments)
           [level, fragments.first]
         end
 
@@ -61,16 +62,20 @@ class Brief::Document::Section
         @cycles += 1
       end
 
-      self.nodes = source.map(&:last)
+      self.nodes = source.map(&:last).compact
 
       nodes.each do |node|
+        # this could be a bad bugfix
+        node = node.first if node.is_a?(Array)
+
         parent = node.css('section, article').first
-        if %w(h1 h2 h3 h4 h5 h6).include?(parent.children.first.name)
+        if parent.children.first && %w(h1 h2 h3 h4 h5 h6).include?(parent.children.first.name)
           parent['data-heading'] = parent.children.first.text
         end
       end
 
-      nodes.map!(&:to_html)
+      nodes.map! {|n| n.to_html rescue nil}
+      nodes.compact!
     end
 
     def maxed_out?