+++ lib/brief/server/handlers/info.rb
@@ -4,17 +4,7 @@ module Brief::Server::Handlers
       request = options.fetch(:request)
       action = options.fetch(:action)
 
-      schema = {}
-      models = {}
-
-      Brief::Model.classes.map do |k|
-        schema[k.type_alias] = k.to_schema
-
-        a = k.type_alias.to_s.pluralize
-        models[a] = briefcase.send(a).map {|m| m.as_json(docs_path: briefcase.docs_path) }
-      end
-
-      [200, {}, {name: briefcase.folder_name.to_s, schema: schema, models: models}]
+      [200, {}, briefcase.as_default]
     end
   end
 end