+++ spec/fixtures/example/models/page.rb
@@ -1,0 +1,11 @@
+class Brief::Page
+  include Brief::Model
+
+  meta do
+    title
+  end
+
+  content do
+    title "h1:first-of-type"
+  end
+end